Added style to LocaleSwitcher component HandleLoginBtn component moved to auth folder
