+++ vite-project/src/pages/auth/HandleLoginBtn.jsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import {Link, useNavigate} from "react-router-dom";
-import {useAuth} from "../providers/AuthProvider.jsx";
-import {AUTH_API_URL} from "../config/constants.jsx";
+import {useAuth} from "../../providers/AuthProvider.jsx";
+import {AUTH_API_URL} from "../../config/constants.jsx";
 
 const HandleLoginBtn = ({type}) => {
     const {isLoggedIn, logout} = useAuth();
@@ -18,7 +18,7 @@ const HandleLoginBtn = ({type}) => {
         })
     }
     const className = type === "desktop"
-        ? "text-sm font-semibold leading-6 text-gray-500 hover:text-gray-900"
+        ? "flex items-center text-sm font-semibold leading-6 text-gray-500 hover:text-gray-900"
         : "-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-500 hover:text-gray-900 hover:bg-gray-50"
 
     if (!isLoggedIn) {
